refactor(ui): extract shared TraceListItem component
Replace the duplicated trace-row markup in TestCaseSummary,
TestCaseTracesCard, and TestEvidenceTraces with a single
shared TraceListItem, and surface trace file size (now
returned by getTestRunCaseTraces and typed on TraceInfo).
